requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
pymysql
pyjwt
flask-cors